Make artifacts Bucket-only and consolidate artifact tests
Drop Dataset sync for artifacts (Datasets is deprecated): remove the
artifact blob patterns from the dataset CommitScheduler allow-list and
from the dataset load path. Bucket upload/download is unchanged.

Consolidate the artifact test suite (8 files -> 6), dropping the
heavily-mocked simulations and keeping critical/low-mock coverage:
- delete test_artifact_remote.py (mocked _predict) and
  test_artifact_download_remote.py (mocked httpx.stream); salvage their
  genuinely-unit and security tests into test_artifact.py and
  test_artifact_server.py
- trim and rename test_artifact_dataset_sync.py ->
  test_artifact_persistence.py, keeping the backend-agnostic parquet
  round-trip / deletion / validation tests
- drop the obsolete dataset allow-patterns test from the bucket suite
