+++ tests/unit/test_artifact_persistence.py
@@ -3,7 +3,6 @@ from pathlib import Path
 
 import pytest
 
-from trackio.dummy_commit_scheduler import DummyCommitScheduler
 from trackio.sqlite_storage import SQLiteStorage
 
 
@@ -300,188 +299,6 @@ def test_artifact_parquet_export_detects_wal_only_change(temp_dir):
 
     assert pq.stat().st_mtime_ns != first_mtime
     assert len(SQLiteStorage._read_parquet_rows(pq)) == 2
-
-
-@pytest.fixture
-def dataset_loader(monkeypatch):
-    from trackio import sqlite_storage as _ss
-
-    def _setup(repo_files, import_from_parquet=None, download=None):
-        monkeypatch.setenv("TRACKIO_DATASET_ID", "user/ds")
-        monkeypatch.setenv("SPACE_REPO_NAME", "user/space")
-        monkeypatch.delenv("TRACKIO_BUCKET_ID", raising=False)
-        monkeypatch.setattr(_ss.SQLiteStorage, "_dataset_import_attempted", False)
-        monkeypatch.setattr(_ss.SQLiteStorage, "_dataset_import_pending", False)
-        monkeypatch.setattr(_ss.SQLiteStorage, "_dataset_remote_synced", False)
-        monkeypatch.setattr(
-            _ss.SQLiteStorage,
-            "get_scheduler",
-            staticmethod(lambda: DummyCommitScheduler()),
-        )
-        monkeypatch.setattr(
-            _ss.SQLiteStorage,
-            "import_from_parquet",
-            staticmethod(import_from_parquet or (lambda: None)),
-        )
-
-        class _FakeApi:
-            def list_repo_files(self, repo_id, repo_type=None):
-                return repo_files
-
-        requested = []
-
-        def _fake_download(dataset_id, file, repo_type=None, local_dir=None):
-            requested.append(file)
-            dest = Path(local_dir) / file
-            dest.parent.mkdir(parents=True, exist_ok=True)
-            dest.write_bytes(b"x")
-            return str(dest)
-
-        monkeypatch.setattr(_ss.hf, "HfApi", lambda: _FakeApi())
-        monkeypatch.setattr(_ss.hf, "hf_hub_download", download or _fake_download)
-        return _ss, requested
-
-    return _setup
-
-
-def test_load_from_dataset_downloads_artifact_blobs(temp_dir, dataset_loader):
-    artifact_blob = "artifacts/proj/blobs/sha256/aa/" + "a" * 64
-
-    _ss, requested = dataset_loader(
-        [
-            "proj.parquet",
-            "proj_artifact_versions.parquet",
-            "media/proj/run/img.png",
-            artifact_blob,
-            "proj.db",
-            "README.md",
-        ]
-    )
-
-    _ss.SQLiteStorage.load_from_dataset()
-
-    assert artifact_blob in requested
-    assert "media/proj/run/img.png" in requested
-    assert "proj.parquet" in requested
-    assert "proj_artifact_versions.parquet" in requested
-    assert "proj.db" not in requested
-    assert "README.md" not in requested
-
-
-def test_load_from_dataset_retries_import_after_transient_failure(
-    temp_dir, dataset_loader
-):
-    """A transient import failure must not permanently block the import. The
-    first call downloads and fails in import_from_parquet, leaving the gate open
-    with the import still pending; the next call re-runs the import even though
-    no new files download, and only then marks the dataset import attempted."""
-    import_calls = []
-
-    def _flaky_import():
-        import_calls.append(1)
-        if len(import_calls) == 1:
-            raise RuntimeError("locked parquet")
-
-    _ss, downloads = dataset_loader(["proj.parquet"], import_from_parquet=_flaky_import)
-
-    _ss.SQLiteStorage.load_from_dataset()
-    assert _ss.SQLiteStorage._dataset_import_attempted is False
-    assert _ss.SQLiteStorage._dataset_import_pending is True
-
-    _ss.SQLiteStorage.load_from_dataset()
-    assert _ss.SQLiteStorage._dataset_import_attempted is True
-    assert _ss.SQLiteStorage._dataset_import_pending is False
-    assert len(import_calls) == 2
-    assert downloads == ["proj.parquet"]
-
-
-def test_load_from_dataset_skips_import_when_nothing_downloaded(
-    temp_dir, dataset_loader
-):
-    """Guards the persistent-storage path: when no new files download, the
-    destructive import_from_parquet must NOT run, so live DB tables are never
-    replaced by a staler parquet snapshot."""
-    (Path(temp_dir) / "proj.parquet").write_bytes(b"x")
-
-    import_calls = []
-
-    def _fail_download(*args, **kwargs):
-        raise AssertionError("must not download an already-present file")
-
-    _ss, _ = dataset_loader(
-        ["proj.parquet"],
-        import_from_parquet=lambda: import_calls.append(1),
-        download=_fail_download,
-    )
-
-    _ss.SQLiteStorage.load_from_dataset()
-    assert _ss.SQLiteStorage._dataset_import_attempted is True
-    assert import_calls == []
-
-
-def test_load_from_dataset_does_not_deadlock_on_reentrant_init_db(
-    temp_dir, monkeypatch
-):
-    import threading
-
-    from trackio import sqlite_storage as _ss
-
-    monkeypatch.setenv("TRACKIO_DATASET_ID", "user/ds")
-    monkeypatch.setenv("SPACE_REPO_NAME", "user/space")
-    monkeypatch.delenv("TRACKIO_BUCKET_ID", raising=False)
-
-    class _RealLockScheduler:
-        def __init__(self):
-            self.lock = threading.Lock()
-
-    sched = _RealLockScheduler()
-    monkeypatch.setattr(_ss.SQLiteStorage, "get_scheduler", staticmethod(lambda: sched))
-    monkeypatch.setattr(_ss.SQLiteStorage, "_dataset_import_attempted", True)
-    monkeypatch.setattr(_ss.SQLiteStorage, "_dataset_import_pending", False)
-    monkeypatch.setattr(_ss.SQLiteStorage, "_dataset_remote_synced", False)
-
-    SQLiteStorage.init_db("proj")
-    SQLiteStorage.log(project="proj", run="train", metrics={"loss": 0.1})
-    SQLiteStorage.export_to_parquet()
-
-    remote = Path(temp_dir) / "_remote"
-    remote.mkdir()
-    moved = []
-    for pq in Path(temp_dir).glob("*.parquet"):
-        pq.rename(remote / pq.name)
-        moved.append(pq.name)
-    db_path = SQLiteStorage.get_project_db_path("proj")
-    for suffix in ("", "-wal", "-shm", "-journal"):
-        sidecar = db_path.with_name(db_path.name + suffix)
-        if sidecar.exists():
-            sidecar.unlink()
-
-    class _FakeApi:
-        def list_repo_files(self, repo_id, repo_type=None):
-            return moved
-
-    def _fake_download(dataset_id, file, repo_type=None, local_dir=None):
-        dest = Path(local_dir) / file
-        dest.parent.mkdir(parents=True, exist_ok=True)
-        dest.write_bytes((remote / file).read_bytes())
-        return str(dest)
-
-    monkeypatch.setattr(_ss.hf, "HfApi", lambda: _FakeApi())
-    monkeypatch.setattr(_ss.hf, "hf_hub_download", _fake_download)
-    monkeypatch.setattr(_ss.SQLiteStorage, "_dataset_import_attempted", False)
-    monkeypatch.setattr(_ss.SQLiteStorage, "_dataset_import_pending", False)
-
-    done = threading.Event()
-
-    def _run():
-        try:
-            SQLiteStorage.load_from_dataset()
-        finally:
-            done.set()
-
-    threading.Thread(target=_run, daemon=True).start()
-    assert done.wait(timeout=10), "load_from_dataset deadlocked on re-entrant init_db"
-    assert "proj" in SQLiteStorage.get_projects()
 
 
 def test_artifact_parquet_tables_match_schema(temp_dir):
